Promote Price component to src/common
