+++ src/common/price/price.scss
@@ -1,0 +1,8 @@
+.price {
+  span:first-child {
+    text-decoration: line-through;
+  }
+  span:last-child {
+    color: red;
+  }
+}
