+++ src/overview/ProductInformation.test.jsx
@@ -45,31 +45,3 @@ it('sends users to the Ratings & Reviews section on click', () => {
   const app = render(<App info={info} questions={[]} related={[]} reviews={[]}/>);
   expect(app.container.querySelector(anchor)).not.toBeNull();
 });
-
-it('displays price and sale in the correct format', () => {
-  // deep copy
-  const style = JSON.parse(JSON.stringify(styles[0]));
-  style.original_price = '123.00';
-  style.sale_price = '124.01';
-
-  render(<ProductInformation product={product} rating={rating} style={style}/>);
-
-  expect(screen.queryByText('$123')).not.toBeNull();
-  expect(screen.queryByText('$124.01')).not.toBeNull();
-});
-
-it('hides the sale price if there is no sale', () => {
-  const style = JSON.parse(JSON.stringify(styles[0]));
-  style.sale_price = '1';
-  const withSale = render(
-    <ProductInformation product={product} rating={rating} style={style}/>
-  );
-  expect(withSale.container.querySelector('.price').childElementCount).toBe(2);
-
-  style.sale_price = null;
-  const withoutSale = render(
-    <ProductInformation product={product} rating={rating} style={style}/>
-  );
-  expect(withoutSale.container.querySelector('.price').childElementCount)
-    .toBeLessThan(2);
-});
