+++ src/overview/overview.scss
@@ -66,15 +66,6 @@
     margin-left: 0.25em;
   }
 
-  .price {
-    span:first-child {
-      text-decoration: line-through;
-    }
-    span:last-child {
-      color: red;
-    }
-  }
-
   .react-share__ShareButton {
     vertical-align: middle;
     margin-left: 0.25em;
