+++ src/related-items-and-comparisons/related-items-and-comparisons.scss
@@ -15,27 +15,7 @@
   }
 
   .price {
-    font-size: 14px;
-    margin-top: 2.5px;
-    margin-left: 5px;
-  }
-
-  .price-container {
-    display: flex;
-  }
-
-  .sale-price {
-    color: red;
-    font-size: 14px;
-    margin-top: 2.5px;
-    margin-left: 5px;
-  }
-
-  .default-price {
-    font-size: 14px;
-    margin-top: 2.5px;
-    margin-left: 5px;
-    text-decoration: line-through;
+    margin: 2.5px 5px;
   }
 
   .category {
